+++ CHANGELOG.md
@@ -9,6 +9,8 @@ You can find its changes [documented below](#083---2023-02-28).
 
 ### Added
 
+- Support for showing native system menu on window with custom title bar for windows platform ([#2362] by [@YouKnow-sys])
+
 ### Changed
 
 ### Deprecated
@@ -783,6 +785,7 @@ Last release without a changelog :(
 [@jaredoconnell]: https://github.com/jaredoconnell
 [@giannissc]: https://github.com/giannissc
 [@newcomb-luke]: https://github.com/newcomb-luke
+[@YouKnow-sys]: https://github.com/YouKnow-sys
 
 [#599]: https://github.com/linebender/druid/pull/599
 [#611]: https://github.com/linebender/druid/pull/611
@@ -1227,6 +1230,7 @@ Last release without a changelog :(
 [#2352]: https://github.com/linebender/druid/pull/2352
 [#2353]: https://github.com/linebender/druid/pull/2353
 [#2356]: https://github.com/linebender/druid/pull/2356
+[#2362]: https://github.com/linebender/druid/pull/2362
 
 [Unreleased]: https://github.com/linebender/druid/compare/v0.8.3...master
 [0.8.3]: https://github.com/linebender/druid/compare/v0.8.2...v0.8.3